pyproject.toml
[project]
name = "gemma4vla"
version = "0.1.0"
description = "Gemma4VLA: Gemma 4 backbone + π0.6-style action expert for universal robot control"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.1.0,<2.11.0",
    "torchvision>=0.16.0,<0.26.0",
    "transformers>=4.50.0",
    "accelerate>=0.30.0",
    "peft>=0.12.0",
    "numpy>=1.24.0",
    "Pillow>=10.0",
    "pyyaml>=6.0",
    "websockets>=12.0",
    "msgpack>=1.0",
    "msgpack-numpy>=0.4.8",
    "einops>=0.7.0",
    "safetensors>=0.4.0",
    "openpi-client",
    "huggingface-hub>=0.20.0",
    "tqdm>=4.60.0",
]

[project.optional-dependencies]
train = [
    "wandb>=0.16.0",
    "datasets>=2.16.0",
]
dev = ["pytest>=7.0", "ruff>=0.1.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gemma4vla"]

[tool.uv.sources]
openpi-client = { git = "https://github.com/Physical-Intelligence/openpi.git", subdirectory = "packages/openpi-client" }

[tool.ruff]
line-length = 120
target-version = "py311"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
